+++ src/app/(main)/feed/page.tsx
@@ -1,13 +1,8 @@
-"use client";
-
-import Chip from "@/components/commons/chip";
-import FeedItem from "@/components/feed";
 import { MainLayout } from "@/components/layout";
-
 import Pencil from "@/assets/icons/pencil.svg";
 import Search from "@/assets/icons/search.svg";
 import TextLogo from "../../../../public/img/logo-text.svg";
-import { useState } from "react";
+import FeedPageList from "@/components/feed/feed-page";
 
 interface Feed {
   feedId: number;
@@ -24,8 +19,6 @@ interface Feed {
 }
 
 export default function FeedPage() {
-  const [filter, setFilter] = useState<"recommended" | "latest">("recommended");
-
   const headerProps = {
     title: "피드",
     leftIcons: [{ icon: <TextLogo />, path: "/" }],
@@ -35,9 +28,7 @@ export default function FeedPage() {
     ],
   };
 
-  // 추천순, 최신순 필터링 추가
-
-  const FEED_DATA: Feed[] = [
+  const feedData: Feed[] = [
     {
       feedId: 1,
       profileImage: "https://via.placeholder.com/40",
@@ -60,6 +51,7 @@ export default function FeedPage() {
       feedId: 2,
       profileImage: "https://via.placeholder.com/40",
       profileName: "김하비",
+
       time: "2024-12-06T21:42:31.279Z",
       feedFiles: [
         "https://fastly.picsum.photos/id/524/400/400.jpg?hmac=msawrWikHM9c5n3cItDml12lRmNkSrzXlRWyTnoIdig",
@@ -77,35 +69,8 @@ export default function FeedPage() {
   ];
 
   return (
-    <>
-      <MainLayout headerProps={headerProps}>
-        <div className="bg-blue-50 w-full overflow-y-hidden">
-          {/* 버튼으로 교체 예정 */}
-          <div className="flex justify-start items-center gap-2 pl-4 bg-white">
-            <Chip label="추천순" onClick={() => setFilter("recommended")} />
-            <Chip label="최신순" onClick={() => setFilter("latest")} />
-          </div>
-
-          {FEED_DATA.map((feed) => (
-            <FeedItem key={feed.feedId} className="w-full">
-              <FeedItem.Profile
-                profileImage={feed.profileImage}
-                profileName={feed.profileName}
-                isFollowing={feed.isFollowing}
-              />
-              <FeedItem.Image feedFiles={feed.feedFiles} />
-              <FeedItem.Content contents={feed.contents} />
-              <FeedItem.HashTags hashTag={feed.hashTag} />
-              <FeedItem.Actions
-                feedId={feed.feedId}
-                likeCount={feed.likeCount}
-                commentCount={feed.commentCount}
-                liked={feed.liked}
-              />
-            </FeedItem>
-          ))}
-        </div>
-      </MainLayout>
-    </>
+    <MainLayout headerProps={headerProps}>
+      <FeedPageList feedData={feedData} />
+    </MainLayout>
   );
 }
